refactor: hover.ts -> docs.ts and no default export
